removed cli_utils, refactored functions
